requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
bottle==0.13.3
bottle-websocket==0.2.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
Eel==0.18.1
et_xmlfile==2.0.0
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.2
idna==3.10
numpy==2.2.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
PyMySQL==1.1.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2024.2
urllib3==2.4.0
zope.event==5.0
zope.interface==7.2
